Moved deprecation warnings out
