+++ kornia/augmentation/functional/__init__.py
@@ -1,9 +1,2 @@
 from .functional import *
 from .functional3d import *
-
-import warnings
-
-
-def __deprecation_warning(name: str, replacement: str):
-    warnings.warn(f"`{name}` is no longer maintained and will be removed from the future versions. "
-                  f"Please use {replacement} instead.")
